Refactor global agents to split into multiple files
