+++ front/lib/api/assistant/global_agents/global_agent_metadata.ts
