+++ front/pages/api/w/[wId]/assistant/global_agents/[aId]/index.ts
@@ -3,7 +3,7 @@ import * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { upsertGlobalAgentSettings } from "@app/lib/api/assistant/global_agents";
+import { upsertGlobalAgentSettings } from "@app/lib/api/assistant/global_agents/global_agents";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { apiError } from "@app/logger/withlogging";
